test(admin): update admin action tests for the 5 dedicated pages
Point every GET/POST test at the item's new page route instead of the
retired app_admin_actions, and split AdminPageTest into one test class per
new controller (AdminUpdateDataTest, AdminUpdateAvailabilitiesTest,
AdminCalculateDataTest, AdminInvalidateDataTest, AdminTriggerPipelineTest),
adding unit test coverage for each new controller class.

Icon/alert assertions that used to count matches across the whole combined
17-item page are rescoped to the specific item's container id, since that
is what each test actually cares about and it no longer depends on which
other items happen to share the page.

RedirectActionsTest drops the tab-activation step (items are always visible
on their own page now) and asserts a redirect back to the item's own page
instead of always app_admin_actions.
