+++ tests/src/Browser/Admin/ForceActionTest.php
@@ -28,7 +28,7 @@ final class ForceActionTest extends AbstractBrowserTestCase
         $user->addAdminRole();
         $this->loginUser($client, $user);
 
-        $client->request('GET', '/fr/istration/actions');
+        $client->request('GET', '/fr/istration/update_data');
 
         // Sentinel wiped out by any real page load (navigation destroys the JS context), unlike
         // the `data-updated-state` attribute which can't distinguish "no navigation happened" from
@@ -56,7 +56,7 @@ final class ForceActionTest extends AbstractBrowserTestCase
         $user->addAdminRole();
         $this->loginUser($client, $user);
 
-        $client->request('GET', '/fr/istration/actions');
+        $client->request('GET', '/fr/istration/update_data');
 
         $button = $client->findElement(WebDriverBy::cssSelector('#update_games_collections_and_dex button.admin-item-cta'));
         $client->executeScript('arguments[0].scrollIntoView({block: "center", inline: "nearest"});', [$button]);
