+++ tests/src/Browser/Admin/ToggleActionsTest.php
@@ -27,7 +27,7 @@ final class ToggleActionsTest extends AbstractBrowserTestCase
         $user->addAdminRole();
         $this->loginUser($client, $user);
 
-        $client->request('GET', '/fr/istration/actions');
+        $client->request('GET', '/fr/istration/update_data');
 
         $this->assertSelectorIsVisible('.admin-item-update_labels .admin-item-current');
         $this->assertSelectorIsNotVisible('.admin-item-update_labels .admin-item-last');
