+++ tests/src/Integration/Controller/Admin/ActionUpdateTest.php
@@ -23,6 +23,19 @@ final class ActionUpdateTest extends WebTestCase
 {
     use TestNavTrait;
 
+    /**
+     * Items sharing the 'update' action verb are split across two dedicated pages
+     * (see AdminUpdateAvailabilitiesController / AdminUpdateDataController); this mirrors
+     * AdminActionController::resolveRedirectRoute()'s own mapping.
+     *
+     * @var list<string>
+     */
+    private const array AVAILABILITIES_ITEMS = [
+        'games_availabilities',
+        'games_shinies_availabilities',
+        'collections_availabilities',
+    ];
+
     #[Test]
     public function adminUpdateLabels(): void
     {
@@ -64,14 +77,14 @@ final class ActionUpdateTest extends WebTestCase
 
         // We don't have to call 'action/update/games_shinies_availabilities' to see error
         // because it's the last action log we looking for
-        $crawler = $client->request('GET', '/fr/istration/actions');
+        $crawler = $client->request('GET', '/fr/istration/update_availabilities');
 
         $this->assertResponseStatusCodeSame(200);
 
-        $this->assertCountFilter($crawler, 0, '.icon-square.bg-success');
-        $this->assertCountFilter($crawler, 1, '.icon-square.bg-danger');
-        $this->assertCountFilter($crawler, 2, '.alert-danger');
-        $this->assertSelectorTextSame('.alert-danger', 'Exception has been thrown for X reason');
+        $this->assertCountFilter($crawler, 0, '#update_games_shinies_availabilities .icon-square.bg-success');
+        $this->assertCountFilter($crawler, 1, '#update_games_shinies_availabilities .icon-square.bg-danger');
+        $this->assertCountFilter($crawler, 1, '#update_games_shinies_availabilities .alert-danger');
+        $this->assertSelectorTextSame('#update_games_shinies_availabilities .alert-danger', 'Exception has been thrown for X reason');
 
         $this->assertConnectedNavBar($crawler);
         $this->assertFrenchLangSwitch($crawler);
@@ -129,20 +142,19 @@ final class ActionUpdateTest extends WebTestCase
         $user->addAdminRole();
         $client->loginUser($user, 'web');
 
-        $crawler = $client->request('GET', '/fr/istration/actions');
+        $crawler = $client->request('GET', '/fr/istration/update_data');
         $form = $crawler->filter('#update_labels form')->form();
         $client->submit($form);
 
         $this->assertResponseStatusCodeSame(302);
         $crawler = $client->followRedirect();
 
-        $this->assertCountFilter($crawler, 1, '.icon-square.bg-success');
-        $this->assertCountFilter($crawler, 0, '.icon-square.bg-warning');
+        $this->assertCountFilter($crawler, 1, '#update_labels .icon-square.bg-success');
+        $this->assertCountFilter($crawler, 0, '#update_labels .icon-square.bg-warning');
 
-        $crawler = $client->request('GET', '/fr/istration/actions');
+        $crawler = $client->request('GET', '/fr/istration/update_data');
 
-        $this->assertCountFilter($crawler, 0, '.icon-square.bg-success');
-        $this->assertCountFilter($crawler, 0, '.icon-square.bg-warning');
+        $this->assertCountFilter($crawler, 0, '#update_labels .icon-square.bg-warning');
     }
 
     private function testAdminUpdate(string $name): void
@@ -153,13 +165,17 @@ final class ActionUpdateTest extends WebTestCase
         $user->addAdminRole();
         $client->loginUser($user, 'web');
 
-        $crawler = $client->request('GET', '/fr/istration/actions');
+        $page = \in_array($name, self::AVAILABILITIES_ITEMS, true)
+            ? '/fr/istration/update_availabilities'
+            : '/fr/istration/update_data';
+
+        $crawler = $client->request('GET', $page);
         $form = $crawler->filter("#update_{$name} form")->form();
         $client->submit($form);
 
         $this->assertResponseStatusCodeSame(302);
         $crawler = $client->followRedirect();
-        $this->assertSame('http://localhost/fr/istration/actions', $client->getRequest()->getUri());
+        $this->assertSame("http://localhost{$page}", $client->getRequest()->getUri());
 
         $this->assertSame(
             'Pokénini Administration',
@@ -171,7 +187,7 @@ final class ActionUpdateTest extends WebTestCase
             $crawler->filter('h1')->text()
         );
 
-        $this->assertCountFilter($crawler, 1, '.icon-square.bg-success');
+        $this->assertCountFilter($crawler, 1, "#update_{$name} .icon-square.bg-success");
 
         $this->assertConnectedNavBar($crawler);
         $this->assertFrenchLangSwitch($crawler);
