+++ tests/src/Unit/Controller/AdminControllerTest.php
@@ -5,20 +5,11 @@ declare(strict_types=1);
 namespace App\Tests\Unit\Controller;
 
 use App\Controller\AdminController;
-use App\DTO\AdminAction;
-use App\Service\Back\GetActionLogsService;
-use App\Service\Back\GetImagePipelineStatusService;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\HttpFoundation\Session\Flash\FlashBag;
-use Symfony\Component\HttpFoundation\Session\FlashBagAwareSessionInterface;
-use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\RouterInterface;
-use Twig\Environment;
 
 /**
  * @internal
@@ -27,14 +18,14 @@ use Twig\Environment;
 final class AdminControllerTest extends TestCase
 {
     #[Test]
-    public function indexRedirectsToActions(): void
+    public function indexRedirectsToUpdateData(): void
     {
         $router = $this->createMock(RouterInterface::class);
         $router
             ->expects($this->once())
             ->method('generate')
-            ->with('app_admin_actions', [])
-            ->willReturn('/fr/istration/actions')
+            ->with('app_admin_update_data', [])
+            ->willReturn('/fr/istration/update_data')
         ;
 
         $container = $this->createMock(ContainerInterface::class);
@@ -45,222 +36,12 @@ final class AdminControllerTest extends TestCase
             ->willReturn($router)
         ;
 
-        $controller = new AdminController(
-            $this->createStub(GetActionLogsService::class),
-            $this->createStub(GetImagePipelineStatusService::class),
-        );
+        $controller = new AdminController();
         $controller->setContainer($container);
 
         $response = $controller->index();
 
-        $this->assertSame('/fr/istration/actions', $response->getTargetUrl());
+        $this->assertSame('/fr/istration/update_data', $response->getTargetUrl());
         $this->assertSame(302, $response->getStatusCode());
-    }
-
-    #[Test]
-    public function adminAction(): void
-    {
-        $adminAction = new AdminAction(
-            'truc',
-            'machin',
-            'ok',
-            'content',
-            '',
-        );
-
-        $flashBag = new FlashBag();
-
-        $session = $this->createMock(SessionInterface::class);
-        $session
-            ->expects($this->once())
-            ->method('get')
-            ->with('admin.action.response.content')
-            ->willReturn($adminAction)
-        ;
-        $session
-            ->expects($this->once())
-            ->method('remove')
-            ->with('admin.action.response.content')
-        ;
-
-        $flashBagSession = $this->createMock(FlashBagAwareSessionInterface::class);
-        $flashBagSession
-            ->expects($this->exactly(3))
-            ->method('getFlashBag')
-            ->willReturn($flashBag)
-        ;
-
-        $requestStack = $this->createMock(RequestStack::class);
-        $requestStack
-            ->expects($this->exactly(4))
-            ->method('getSession')
-            ->willReturnOnConsecutiveCalls(
-                $session,
-                $flashBagSession,
-                $flashBagSession,
-                $flashBagSession,
-            )
-        ;
-
-        $twig = $this->createMock(Environment::class);
-        $twig
-            ->expects($this->once())
-            ->method('render')
-            ->with(
-                'Admin/actions.html.twig',
-                [
-                    'actionLogsData' => [],
-                    'imagePipelineStatus' => null,
-                ]
-            )
-            ->willReturn('<html></html>')
-        ;
-
-        $container = $this->createMock(ContainerInterface::class);
-        $container
-            ->expects($this->exactly(4))
-            ->method('get')
-            ->willReturnOnConsecutiveCalls(
-                $requestStack,
-                $requestStack,
-                $requestStack,
-                $twig,
-            )
-        ;
-        $container
-            ->expects($this->once())
-            ->method('has')
-            ->with('twig')
-            ->willReturn(true)
-        ;
-
-        $controller = $this->getController();
-        $controller->setContainer($container);
-
-        $controller->actions($requestStack, new Request());
-
-        $this->assertEquals(
-            [
-                'action' => ['truc'],
-                'item' => ['machin'],
-                'state' => ['ok'],
-            ],
-            $flashBag->all()
-        );
-    }
-
-    #[Test]
-    public function adminActionError(): void
-    {
-        $adminAction = new AdminAction(
-            'truc',
-            'machin',
-            'ok',
-            'content',
-            'error',
-        );
-
-        $flashBag = new FlashBag();
-
-        $session = $this->createMock(SessionInterface::class);
-        $session
-            ->expects($this->once())
-            ->method('get')
-            ->with('admin.action.response.content')
-            ->willReturn($adminAction)
-        ;
-        $session
-            ->expects($this->once())
-            ->method('remove')
-            ->with('admin.action.response.content')
-        ;
-
-        $flashBagSession = $this->createMock(FlashBagAwareSessionInterface::class);
-        $flashBagSession
-            ->expects($this->exactly(4))
-            ->method('getFlashBag')
-            ->willReturn($flashBag)
-        ;
-
-        $requestStack = $this->createMock(RequestStack::class);
-        $requestStack
-            ->expects($this->exactly(5))
-            ->method('getSession')
-            ->willReturnOnConsecutiveCalls(
-                $session,
-                $flashBagSession,
-                $flashBagSession,
-                $flashBagSession,
-                $flashBagSession,
-            )
-        ;
-
-        $twig = $this->createMock(Environment::class);
-        $twig
-            ->expects($this->once())
-            ->method('render')
-            ->with(
-                'Admin/actions.html.twig',
-                [
-                    'actionLogsData' => [],
-                    'imagePipelineStatus' => null,
-                ]
-            )
-            ->willReturn('<html></html>')
-        ;
-
-        $container = $this->createMock(ContainerInterface::class);
-        $container
-            ->expects($this->exactly(5))
-            ->method('get')
-            ->willReturnOnConsecutiveCalls(
-                $requestStack,
-                $requestStack,
-                $requestStack,
-                $requestStack,
-                $twig,
-            )
-        ;
-        $container
-            ->expects($this->once())
-            ->method('has')
-            ->with('twig')
-            ->willReturn(true)
-        ;
-
-        $controller = $this->getController();
-        $controller->setContainer($container);
-
-        $controller->actions($requestStack, new Request());
-
-        $this->assertEquals(
-            [
-                'action' => ['truc'],
-                'item' => ['machin'],
-                'state' => ['ok'],
-                'error' => ['error'],
-            ],
-            $flashBag->all()
-        );
-    }
-
-    private function getController(): AdminController
-    {
-        $getActionLogsService = $this->createMock(GetActionLogsService::class);
-        $getActionLogsService
-            ->expects($this->once())
-            ->method('get')
-            ->willReturn([])
-        ;
-
-        $getImagePipelineStatusService = $this->createMock(GetImagePipelineStatusService::class);
-        $getImagePipelineStatusService
-            ->expects($this->once())
-            ->method('get')
-            ->with(false)
-            ->willReturn(null)
-        ;
-
-        return new AdminController($getActionLogsService, $getImagePipelineStatusService);
     }
 }
